Modified the datatype of stucture members

+++ stats.h
@@ -1,7 +1,7 @@
 struct Stats {
-  int average;
-  int min;
-  int max;
+  float average;
+  float min;
+  float max;
 };
 struct Stats compute_statistics(const float* numberset, int setlength);
